+++ src/App.js
@@ -12,7 +12,7 @@ import CookiePolicy from './pages/footer/CookiePolicy';
 import TermsOfUse from './pages/footer/TermsofUse';
 import FAQ from './pages/footer/FAQ';
 import Chat from "./pages/chat/Chat";
-import PerfumeDatabase from './pages/perfumedatabase/PerfumeDatabase';
+import PerfumeList from './pages/booklist/PerfumeList';
 
 
 function App() {
@@ -35,8 +35,8 @@ function App() {
         <Route path="/FAQ" element={<Layout />}>
           <Route index element={<FAQ/>} />
         </Route>
-        <Route path="/PerfumeDatabase" element={<Layout />}>
-          <Route index element={<PerfumeDatabase/>} />
+        <Route path="/PerfumeList" element={<Layout />}>
+          <Route index element={<PerfumeList/>} />
           {/* 다른 페이지도 Layout 내에서 정의 */}
         </Route>
           <Route path="/login" element={<Login/>}/>
