+++ src/components/commons/footer/Footer.js
@@ -46,7 +46,7 @@ const Footer = () => {
                                 <span className="separator">|</span>
                                 <a href="/fragrance">향료 알아가기</a>
                                 <span className="separator">|</span>
-                                <a href="/PerfumeDatabase">향수 알아가기</a>
+                                <a href="/PerfumeList">향수 알아가기</a>
                                 <span className="separator">|</span>
                                 <a href="/recommend">맞춤 향수 추천</a>
                                 <span className="separator">|</span>
